@@ -15,11 +15,11 @@
     "js-dev": "npx nodemon build/app.js"
   },
   "dependencies": {
-    "@ilovepdf/ilovepdf-js": "^0.1.0",
-    "@ilovepdf/ilovepdf-nodejs": "^0.1.1",
+    "@ilovepdf/ilovepdf-js": "^0.2.0",
+    "@ilovepdf/ilovepdf-nodejs": "^0.2.1",
     "@techie04/xpresser-mailer": "^0.2.1",
-    "@xpresser/auth": "^0.2.0",
-    "@xpresser/env": "^1.0.3",
+    "@xpresser/auth": "^1.0.2",
+    "@xpresser/env": "^2.1.1",
     "@xpresser/file-uploader": "^2.0.4",
     "@xpresser/session": "^1.4.2",
     "@xpresser/xpress-mongo": "^1.1.8",
@@ -28,26 +28,26 @@
     "cookie-parser": "^1.4.5",
     "cron-time-generator": "^1.2.0",
     "express-session": "^1.17.2",
-    "jsonwebtoken": "^8.5.1",
+    "jsonwebtoken": "^9.0.0",
     "mjml": "^4.11.0",
     "moment": "^2.29.1",
-    "passport": "^0.5.0",
+    "passport": "^0.6.0",
     "passport-google-oauth": "^2.0.0",
     "prettier": "^2.2.1",
     "promise-polyfill": "^8.2.1",
-    "sharp": "^0.29.3",
+    "sharp": "^0.32.1",
     "slugify": "^1.6.2",
+    "typescript": "^5.1.3",
     "xpress-mongo": "^2.1.0",
-    "xpresser": "^0.26.11"
+    "xpresser": "^0.31.1"
   },
   "devDependencies": {
     "@types/mjml": "^4.7.0",
-    "@types/node": "^17.0.12",
-    "@types/sharp": "^0.29.4",
+    "@types/node": "^20.3.0",
+    "@types/sharp": "^0.32.0",
     "@xpresser/ngrok": "^2.2.1",
     "nodemon": "^2.0.15",
-    "ts-node-dev": "^1.1.6",
-    "typescript": "^4.5.3",
+    "ts-node-dev": "^2.0.0",
     "xjs-cli": "^1.1.3"
   }
 }
